Post users.profile.set as a JSON body
It's better to post nested objects as JSON rather than url encoding
them, and this fixes some type errors.
